Suggested solution to late UI by Project Mentor
I replaced the sequential await statements on Line 190 and 192 with the following chain:

  const allData = await fetch('http://localhost:8081/all');
    .then(response => response.json());
I put this in place of Line 190 and removed Line 192.

This forces the response.json() to be called only after fetch finishes. The remaining code executes after the await call is done. Seems to fix the behavior for me. Now the UI updates only after the full API response is available as JSON.